books-scraper-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
# lxml removido - usaremos html.parser